requirements.txt
aiogram
tortoise-orm
tortoise-orm[asyncmy]
aerich
cryptography
redis
aioshutil
httpx
python-dotenv
aiogram-dialog
dataclasses-json
PyJWT